Remove the my_font creation from the top of programs
I mistakenly put it there because I was referring to it in correctly
in another program.  my_font is not stored as an object but rather a
font object which can be referred to as such family='my_font'.
This is the same as loading the LibreBaskerville font in this case.